#43 btnのディレクトリの方が自然なので移動。 また、親のやってることを知らないシンプルな責務にした。
